// projects/ngx-location-viewer/src/lib/types/button-actions.enum.ts
export enum ButtonActions {
    none = '',
    select = 'select',
    selectPolygon = 'select-polygon',
    selectRectangle = 'select-rectangle',
    selectZone = 'select-zone',
    measurement = 'measurement',
    distance = 'measurement-distance',
    area = 'measurement-area',
    whatishere = 'whatishere'
}